Small refactoring & Better Re-organization
Small refactoring of the code. Re-organization of the files and folders. Now they are organized based on their nature
