+++ client/src/components/App.js
@@ -3,15 +3,15 @@ import {BrowserRouter as Router, Switch, Route} from 'react-router-dom'
 import PropTypes from 'prop-types'
 import UserRoute from '../routes/UserRoute'
 import GuestRoute from '../routes/GuestRoute'
-import Header from './Header'
-import LoginPage from './LoginPage'
-import Dashboard from './Dashboard'
-import Statistics from './Statistics'
-import Logs from './Logs'
-import InsertUser from './InsertUserPage'
-import ModifyUser from './ModifyUser'
-import DeleteUser from './DeleteUser'
-import NotFound from './NotFound'
+import Header from './header/Header'
+import LoginPage from './pages/LoginPage'
+import Dashboard from './pages/Dashboard'
+import Statistics from './pages/Statistics'
+import Logs from './pages/Logs'
+import InsertUser from './pages/InsertUserPage'
+import ModifyUser from './pages/ModifyUser'
+import DeleteUser from './pages/DeleteUser'
+import NotFound from './pages/NotFound'
 
 const App = ({location}) => (
   <div>
