+++ client/src/components/pages/LoginPage.js
@@ -1,8 +1,8 @@
 import React, {Component} from 'react'
 import PropTypes from 'prop-types'
 import {connect} from 'react-redux'
-import LoginForm from './LoginForm'
-import {login} from '../actions/auth'
+import LoginForm from '../forms/LoginForm'
+import {login} from '../../actions/auth'
 
 class LoginPage extends Component {
   constructor (props) {
